+++ src/main/java/de/samply/notification/OperationType.java
@@ -11,5 +11,4 @@ public enum OperationType {
     EDIT_QUERY,
     ADD_DOCUMENT,
     EDIT_PROJECT
-
 }