+++ src/main/java/de/samply/token/dto/OpalProjectStatus.java
@@ -1,0 +1,7 @@
+package de.samply.token.dto;
+
+public enum OpalProjectStatus {
+    CREATED,
+    WITH_DATA,
+    NOT_FOUND
+}