+++ src/main/java/de/samply/token/dto/OpalTokenStatus.java
@@ -1,0 +1,7 @@
+package de.samply.token.dto;
+
+public enum OpalTokenStatus {
+    CREATED,
+    EXPIRED,
+    NOT_FOUND
+}