+++ src/main/java/de/samply/user/UserService.java
@@ -4,8 +4,11 @@ import de.samply.db.model.*;
 import de.samply.db.repository.*;
 import de.samply.notification.NotificationService;
 import de.samply.notification.OperationType;
+import de.samply.project.ProjectType;
 import de.samply.project.state.UserProjectState;
 import de.samply.security.SessionUser;
+import de.samply.token.TokenManagerService;
+import de.samply.token.TokenManagerServiceException;
 import de.samply.user.roles.ProjectRole;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.stereotype.Service;
@@ -22,6 +25,7 @@ public class UserService {
     private final ProjectBridgeheadUserRepository projectBridgeheadUserRepository;
     private final ProjectRepository projectRepository;
     private final ProjectBridgeheadRepository projectBridgeheadRepository;
+    private final TokenManagerService tokenManagerService;
     private final SessionUser sessionUser;
 
     public UserService(NotificationService notificationService,
@@ -30,6 +34,7 @@ public class UserService {
                        ProjectBridgeheadUserRepository projectBridgeheadUserRepository,
                        ProjectRepository projectRepository,
                        ProjectBridgeheadRepository projectBridgeheadRepository,
+                       TokenManagerService tokenManagerService,
                        SessionUser sessionUser) {
         this.notificationService = notificationService;
         this.bridgeheadAdminUserRepository = bridgeheadAdminUserRepository;
@@ -37,6 +42,7 @@ public class UserService {
         this.projectBridgeheadUserRepository = projectBridgeheadUserRepository;
         this.projectRepository = projectRepository;
         this.projectBridgeheadRepository = projectBridgeheadRepository;
+        this.tokenManagerService = tokenManagerService;
         this.sessionUser = sessionUser;
     }
 
@@ -84,7 +90,7 @@ public class UserService {
         return result;
     }
 
-    public void setProjectBridgheadUserWithRole(@NotNull String email, @NotNull String projectCode, @NotNull String bridgehead, @NotNull ProjectRole projectRole) throws UserServiceException {
+    public void setProjectBridgheadUserWithRoleAndGenerateTokensIfDataShield(@NotNull String email, @NotNull String projectCode, @NotNull String bridgehead, @NotNull ProjectRole projectRole) throws UserServiceException {
         Optional<Project> project = this.projectRepository.findByCode(projectCode);
         if (project.isEmpty()) {
             throw new UserServiceException("Project " + projectCode + " not found");
@@ -104,6 +110,18 @@ public class UserService {
             this.projectBridgeheadUserRepository.save(projectBridgeheadUser);
             this.notificationService.createNotification(projectCode, bridgehead, email, OperationType.ASSIGN_USER_TO_PROJECT,
                     "Set role " + projectRole + " to user", null, null);
+        }
+        if (project.get().getType() == ProjectType.DATASHIELD) {
+            generateTokensAndProjectsInOpal(projectCode, bridgehead, projectBridgeheadUserOptional.get());
+        }
+    }
+
+    private void generateTokensAndProjectsInOpal(@NotNull String projectCode, @NotNull String bridgehead, @NotNull ProjectBridgeheadUser projectBridgeheadUser) throws UserServiceException {
+        try {
+            tokenManagerService.generateTokensAndProjectsInOpal(projectCode, bridgehead,
+                    () -> projectBridgeheadUserRepository.delete(projectBridgeheadUser));
+        } catch (TokenManagerServiceException e) {
+            throw new UserServiceException(e);
         }
     }